fix: move user input cleanup to after_build
