+++ src/Drupal/FormValidatorDrupal.php
@@ -104,67 +104,9 @@ final class FormValidatorDrupal {
    */
   public static function typeCastRecursive(array $element, FormStateInterface $form_state, object $schema): void {
     $data = $form_state->getValue($element['#parents']);
-    $data = static::cleanUserInput($data);
+    $data = UserInputCleaner::cleanUserInput($data);
     $data = RecursiveTypeCaster::recursiveTypeRefinements($data, $schema);
     $form_state->setValueForElement($element, $data);
-  }
-
-  /**
-   * Cleans the input.
-   *
-   * @param mixed $data
-   *   The input before cleaning.
-   *
-   * @return array|mixed
-   *   The clean input.
-   */
-  private static function cleanUserInput(mixed $data) {
-    if (!is_array($data)) {
-      return $data;
-    }
-    if (array_is_list($data)) {
-      return static::arrayTrim($data);
-    }
-    foreach ($data as $key => $datum) {
-      // If the data was left empty in a fieldset, remove it.
-      if ($datum === '') {
-        unset($data[$key]);
-        continue;
-      }
-      if ($key === 'add_more' && $datum instanceof MarkupInterface) {
-        unset($data[$key]);
-        continue;
-      }
-      $clean_datum = static::cleanUserInput($datum);
-      if (is_null($clean_datum)) {
-        unset($data[$key]);
-        continue;
-      }
-      $data[$key] = $clean_datum;
-    }
-    return $data;
-  }
-
-  /**
-   * Trims an array of values with empty strings.
-   *
-   * @param array $data
-   *   The data to trim.
-   *
-   * @return array
-   *   The trimmed data.
-   */
-  private static function arrayTrim(array $data): array {
-    if (!array_is_list($data)) {
-      return $data;
-    }
-    $last_non_empty = -1;
-    foreach ($data as $index => $value) {
-      if ($value !== '') {
-        $last_non_empty = $index;
-      }
-    }
-    return array_slice($data, 0, $last_non_empty + 1);
   }
 
 }
